+++ examples/demo_advanced_workflow_integration.py
@@ -31,7 +31,7 @@ from core.advanced_workflow_engine import (
 from core.agent_manager import AgentManager, AgentStatus, AgentCapability
 from core.config_manager import ConfigManager
 from core.contract_manager import ContractManager, ContractPriority
-from core.performance_tracker import PerformanceTracker
+from core.performance_monitor import PerformanceMonitor
 
 # Configure logging
 logging.basicConfig(
@@ -56,7 +56,7 @@ class AdvancedWorkflowIntegrationDemo:
         self.config_manager = ConfigManager(self.config_dir)
         self.agent_manager = AgentManager(self.agent_dir)
         self.contract_manager = ContractManager(self.agent_manager, self.config_manager)
-        self.performance_tracker = PerformanceTracker(
+        self.performance_tracker = PerformanceMonitor(
             self.agent_manager, self.config_manager
         )
